requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
dataclasses==0.6
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.7.0
django-heroku==0.0.0
djangorestframework==3.12.4
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
pycodestyle==2.7.0
psycopg2==2.8.6
pymongo==3.11.4
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
